Enum linera_alloy_syn_solidity::SubDenomination

pub enum SubDenomination {
    Wei(wei),
    Gwei(gwei),
    Ether(ether),
    Seconds(seconds),
    Minutes(minutes),
    Hours(hours),
    Days(days),
    Weeks(weeks),
    Years(years),
}

A sub-denomination suffix for a number literal.

Variants

Wei(wei)

wei

Gwei(gwei)

gwei

Ether(ether)

ether

Seconds(seconds)

seconds

Minutes(minutes)

minutes

Hours(hours)

hours

Days(days)

days

Weeks(weeks)

weeks

Years(years)

years

Implementations

impl SubDenomination

pub fn new_wei(span: Span) -> Self

Creates a new Wei keyword with the given span.

pub fn new_gwei(span: Span) -> Self

Creates a new Gwei keyword with the given span.

pub fn new_ether(span: Span) -> Self

Creates a new Ether keyword with the given span.

pub fn new_seconds(span: Span) -> Self

Creates a new Seconds keyword with the given span.

pub fn new_minutes(span: Span) -> Self

Creates a new Minutes keyword with the given span.

pub fn new_hours(span: Span) -> Self

Creates a new Hours keyword with the given span.

pub fn new_days(span: Span) -> Self

Creates a new Days keyword with the given span.

pub fn new_weeks(span: Span) -> Self

Creates a new Weeks keyword with the given span.

pub fn new_years(span: Span) -> Self

Creates a new Years keyword with the given span.

pub fn parse_opt(input: ParseStream<'_>) -> Result<Option<Self>>

pub fn peek(lookahead: &Lookahead1<'_>) -> bool

pub const fn as_str(self) -> &'static str

pub const fn as_debug_str(self) -> &'static str

pub const fn is_wei(self) -> bool

Returns true if self matches Self::Wei.

pub const fn is_gwei(self) -> bool

Returns true if self matches Self::Gwei.

pub const fn is_ether(self) -> bool

Returns true if self matches Self::Ether.

pub const fn is_seconds(self) -> bool

Returns true if self matches Self::Seconds.

pub const fn is_minutes(self) -> bool

Returns true if self matches Self::Minutes.

pub const fn is_hours(self) -> bool

Returns true if self matches Self::Hours.

pub const fn is_days(self) -> bool

Returns true if self matches Self::Days.

pub const fn is_weeks(self) -> bool

Returns true if self matches Self::Weeks.

pub const fn is_years(self) -> bool

Returns true if self matches Self::Years.