Enum linera_alloy_syn_solidity::UserDefinableOperator

pub enum UserDefinableOperator {

    BitAnd(And),
    BitNot(Tilde),
    BitOr(Or),
    BitXor(Caret),
    Add(Plus),
    Div(Slash),
    Rem(Percent),
    Mul(Star),
    Sub(Minus),
    Eq(EqEq),
    Ge(Ge),
    Gt(Gt),
    Le(Le),
    Lt(Lt),
    Ne(Ne),

}

A user-definable operator: +, *, |, etc.

Solidity reference: https://docs.soliditylang.org/en/latest/grammar.html#a4.SolidityParser.userDefinableOperator

Variants

BitAnd(And)

&

BitNot(Tilde)

~

BitOr(Or)

|

BitXor(Caret)

^

Add(Plus)

+

Div(Slash)

/

Rem(Percent)

%

Mul(Star)

*

Sub(Minus)

-

Eq(EqEq)

==

Ge(Ge)

>=

Gt(Gt)

>

Le(Le)

<=

Lt(Lt)

<

Ne(Ne)

!=

Implementations

impl UserDefinableOperator

pub fn new_bit_and(span: Span) -> Self

Creates a new BitAnd operator with the given span.

pub fn new_bit_not(span: Span) -> Self

Creates a new BitNot operator with the given span.

pub fn new_bit_or(span: Span) -> Self

Creates a new BitOr operator with the given span.

pub fn new_bit_xor(span: Span) -> Self

Creates a new BitXor operator with the given span.

pub fn new_add(span: Span) -> Self

Creates a new Add operator with the given span.

pub fn new_div(span: Span) -> Self

Creates a new Div operator with the given span.

pub fn new_rem(span: Span) -> Self

Creates a new Rem operator with the given span.

pub fn new_mul(span: Span) -> Self

Creates a new Mul operator with the given span.

pub fn new_sub(span: Span) -> Self

Creates a new Sub operator with the given span.

pub fn new_eq(span: Span) -> Self

Creates a new Eq operator with the given span.

pub fn new_ge(span: Span) -> Self

Creates a new Ge operator with the given span.

pub fn new_gt(span: Span) -> Self

Creates a new Gt operator with the given span.

pub fn new_le(span: Span) -> Self

Creates a new Le operator with the given span.

pub fn new_lt(span: Span) -> Self

Creates a new Lt operator with the given span.

pub fn new_ne(span: Span) -> Self

Creates a new Ne operator with the given span.

pub fn peek(input: ParseStream<'_>, lookahead: &Lookahead1<'_>) -> bool

pub const fn as_str(self) -> &'static str

pub const fn as_debug_str(self) -> &'static str

pub const fn is_bit_and(self) -> bool

Returns true if self matches Self::BitAnd.

pub const fn is_bit_not(self) -> bool

Returns true if self matches Self::BitNot.

pub const fn is_bit_or(self) -> bool

Returns true if self matches Self::BitOr.

pub const fn is_bit_xor(self) -> bool

Returns true if self matches Self::BitXor.

pub const fn is_add(self) -> bool

Returns true if self matches Self::Add.

pub const fn is_div(self) -> bool

Returns true if self matches Self::Div.

pub const fn is_rem(self) -> bool

Returns true if self matches Self::Rem.

pub const fn is_mul(self) -> bool

Returns true if self matches Self::Mul.

pub const fn is_sub(self) -> bool

Returns true if self matches Self::Sub.

pub const fn is_eq(self) -> bool

Returns true if self matches Self::Eq.

pub const fn is_ge(self) -> bool

Returns true if self matches Self::Ge.

pub const fn is_gt(self) -> bool

Returns true if self matches Self::Gt.

pub const fn is_le(self) -> bool

Returns true if self matches Self::Le.

pub const fn is_lt(self) -> bool

Returns true if self matches Self::Lt.

pub const fn is_ne(self) -> bool

Returns true if self matches Self::Ne.